Backport r178340 from trunk:
network_io/unix/sendrecv.c: Deal with EAGAIN after poll().

Following apr_wait_for_io_or_timeout(), we were doing I/O and not
dealing with EAGAIN.  It is legal for poll() to indicate that a socket
is available for writing and then for write() to fail with EAGAIN if
the state of the socket changed in between the poll() and write()
calls.  This only seems to actually happen on Mac OS 10.4 (Darwin 8).

Rather than trying write() only once, if we get an EAGAIN, continue to
call apr_wait_for_io_or_timeout() and try writing.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/0.9.x@178842 13f79535-47bb-0310-9956-ffa450edef68