Merge r1666411 from trunk.
skiplist: restore back add-if-not-exist semantic to apr_skiplist_insert().

Show in the tests how to achieve add-next-to-any-existing semantic by emulating
apr_skiplist_add() (which is 1.6+ only) with a generic compare function.
This is not very nice (brainer, not thread-safe and with an opaque struct abuse)
but generic hence easing code sync with upper branches...


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/1.5.x@1666446 13f79535-47bb-0310-9956-ffa450edef68